+++ CHANGES.md
@@ -7,6 +7,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [0.11.0] - 2021-06-12
+
 ### h3ron crate
 
 #### Added